Fix issue with concat not adding cycRoadType
The microtrips / concat pair, concat was not
perfectly recreating the input from a call to
microtrips. Adding the optional name argument
allows for all keys to be present that were in
the original get_cycle_dict call.

Also, the cycRoadType was not being re-added
back to the cycle within the concat code. This
is now fixed and the code refactored to remove
duplication of key specification.