Fix issues with rate-key interpolation and enhance tests
Note: the strategy used here is to get
the step average values of the rate
variable at the NEW resampled cadence
and associated those sampled values with
the mid-step time and then interpolate
to the desired sample times. This may
result in starting or ending speeds
that are non-zero (assuming cycMps is
in the rate_keys). However, this most
correctly preserves the interpretation
of the data.