Add tests for new np_argmax fn
Ensure that np_argmax is robust for empty
arrays. NOTE: this still may not be the
greatest way to handle an empty array
as the index that is returned (i.e., 0 for an
empty array), cannot be used to lookup
that array. Possibly, we may want to return
as an Option type?