Add stub for a pure Rust vehicle loading function
This was added to have an in-Rust
method of testing that post_init
was working. More could be added
to actually parse a vehicle db file
entry and load a vehicle from disk
without Python. Leaving this in as
a proposed stub but it is in a
single commit so easy to revert
if we don't want to entertain it now.