Extract initail part of sim_drive_walk into init_for_step
This allows for FASTSim analysts to be able to more-easily set up to run sim_drive_step(...) in a loop. init_for_step is the same as the pre-amble for sim_drive_step but also allows for init_soc to be optional. If not provided, init_soc will default to veh.max_soc. The Rust version mimics the Python version.