Implement collision-detection and delay index on python side
The delay index is for the case when coasting to a stop but the lead vehicle departs prior to finishing coast. With the delay index, we go through with the stop but we immediately depart and follow the same trace as the lead vehicle when it departed. An eco-approach / eco-coast vehicle, therefore, will only close the gap by eliminating stopped-time. Therefore, all stops in a cycle will be hit but the dwell time will be small if the coasting vehicle is behind the lead vehicle