Fix issues with coast delay implementation
For eco-coast, we now have the concept of a "coast delay". The coast delay basically is an offset forward or backwards on the reference cycle for where we should pull speed vs time values. This is used to orient the vehicle after coast to either wait for the shadow trace if ahead or immediately start driving again (i.e., minimize stops to a single time-step) if behind.