Correct issue with collision prevention
Could not re-rendezvous with the lead vehicle in some cases (e.g., lead vehicle brakes very sharply from high speed). To correct this, we wrap the rendezvous code in a "full-brake steps" for-loop that introduces some full-on braking prior to trying to find a constant jerk rendezvous solution.

This commit also turns off the majority of print-level debugging.