Fix average_grade_over_range to handle distances, elevations, and grade lookup/calculations correctly
There is an issue with using dist_m to calculate the distance at each sample point. dist_v2_m corrects that issue. This switches to using dist_v2_m and also uses dist_v2_m to calculate the change in elevation from start.

Also, for grade lookup where delta_distance is <= 0.0, we use a single point lookup that would be equivalent to scipy.interp1d with kind = 'previous'. That is, the grade at sample i is assumed to be constant over the step from sample i to sample i+1. This now returns that value.