Add blend factor to eco-cruise
The blend factor is a value between 0 and 1 that blends between using the average microtrip speed and the average microtrip speed while moving only (0 being the former, 1 being the latter, and in the middle, a linear blend between the two).

This was created as a potential optimization variable to explore.