Fix issue when running demos from within venv.
The code to run the demo tests starts a subprocess.
However, that python version may not have the required
libraries installed if FASTSim had been installed to
a virtual environment. This change makes it such that we
use the current python executable to also run the demo
tests. If that executable doesn't resolve (empty string or
None), then we default to just "python" and hope it is on
the path (and set up with FASTSim).