Fix issue with .github\workflows\wheels.yml
The `wheels.yml` file did not receive the required
updates and changes needed for building with Mac OS
for Apple Silicon AND Apple Intel. These changes have
now been copied over such that the `wheels.yml` file is
essentially the same as the first part of the `release.yml`
file. We may be able to eliminate the `wheels.yml` workflow
in the future although it is helpful as a check to confirm
that the `release.yml` file (at least, the "build" section)
will run without issue.