Add (temporary) zeroing of aux loads.
With pwr_aux_base set to 0 kW, we can actually run the HEV label test without panic to the end. This shows that there must be a discrepancy in how aux loads are being handled at the "max power" phase vs the backward solution phase. This seems to only happen for the HEV.