Enable hev to run without panic.
This checks the electric machine for whether
the maximum output is being requested (either
for tractive effort or regen). If so, the same
efficiency value used when calculating the max
possible effort is reused as the efficiency value.
This prevents slight lookup discrepancies and
allows the backward path to match the forward
path for the powertrain.