Fix issue with speeds and add tolerance.
The speed issue was that we were pulling in
the achieved speed in m/s instead of mph. For
the tolerance, we were comparing f64 values
for direct equality. Now we add a bit of tolerance
in to see that the absolute difference (or the
absolute difference fraction, (a - b)/b) is within
tolerance. We don't expect that individual
tolerances should need to be exposed for this
function because the fastsim 2 and 3 results
should be quite close. We just need to account
for some floating point error.